Added new entities to DisguiseType Changed entity Ids to reflect new entity Ids in Spigot 1.9 Cleaned up code

+++ src/main/java/me/libraryaddict/disguise/disguisetypes/watchers/AgeableWatcher.java
@@ -17,7 +17,7 @@ public class AgeableWatcher extends LivingWatcher {
     }
 
     public boolean isBaby() {
-        return ((Byte) getValue(12, (byte) 0)).intValue() < 0;
+        return ((byte) getValue(12, (byte) 0)) < 0;
     }
 
     public void setAdult() {
